requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
dj-rest-auth==5.0.2
Django==4.1.13
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
idna==3.6
mysql-connector-python==8.3.0
mysqlclient==2.2.4
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
